Extract class generation code, refactor tests to separate modules
